Change evaluation metric to BER (or PER)
    But, most of parsers uses float -- not double -- so it should be
    changed too.

* Read and parse double correctly

* Convert BER to PER correctly

* BER means BIT error rate, so PER = BER * 16384 * 8

* Use 5 BERs, 1e-12, 5e-12, 1e-11, 5e-11, and 1e-10

* Add 3.32e-11, an average of 5 BERs

* Switch to 2-level bad page table

* Add debug info about actual good pages

+++ ftl/ftl.cc
@@ -34,7 +34,7 @@ FTL::FTL(ConfigReader &c, DRAM::AbstractDRAM *d) : conf(c), pDRAM(d) {
   param.totalPhysicalBlocks = palparam->superBlock;
   param.totalLogicalBlocks =
       palparam->superBlock *
-      (1 - conf.readFloat(CONFIG_FTL, FTL_OVERPROVISION_RATIO));
+      (1 - conf.readDouble(CONFIG_FTL, FTL_OVERPROVISION_RATIO));
   param.pagesInBlock = palparam->page;
   param.pageSize = palparam->superPageSize;
   param.ioUnitInPage = palparam->pageInSuperPage;
